refactor: group animation components
