+++ components/animations/animated-object.tsx
