+++ components/animations/animated-text.tsx
