+++ components/animations/parallax-image.tsx
