+++ components/animations/parallax-object.tsx
