+++ components/footer/socials.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import ParallaxObject from "../parallax-object";
+import ParallaxObject from "../animations/parallax-object";
 
 export default function Socials() {
   const socials = [
